requirements_mac.txt
clize==5.0.0
gdown==4.7.1
imageio==2.28.0
mediapipe_silicon==0.9.2.1
numpy==1.23.5
onnxruntime_silicon==1.13.1
opencv_contrib_python==4.7.0.72
opencv_python_headless==4.7.0.72
tqdm==4.65.0
